# Conflicts:
#   app/Http/Controllers/QuestionController.php

+++ app/Http/Controllers/AnswerController.php
@@ -6,17 +6,16 @@ use App\Models\Answer;
 use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
-class AnswerController extends Controller
-{
+class AnswerController extends Controller {
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(int $id)
     {
+        $answer = Answer::findOrFail($id);
+        $this->authorize('edit', $answer);
         try {
-            return view('answer.edit', [
-                'answer' => Answer::where('id', $id)->firstOrFail()
-            ]);
+            return view('answer.edit', ['answer' => $asnwer]);
         }
         catch (ModelNotFoundException $e) {
             return "Answer not found.";
@@ -28,8 +27,9 @@ class AnswerController extends Controller
      */
     public function update(Request $request, int $id)
     {
+        $answer = Answer::findOrFail($id);
+        $this->authorize('update', $answer);
         try {
-            $answer = Answer::where('id', $id)->firstOrFail();
             $answer->content = $request->input('content');
             $answer->save();
             return redirect('questions/' . $answer->id_question);
@@ -44,8 +44,9 @@ class AnswerController extends Controller
      */
     public function destroy(int $id)
     {
+        $answer = Answer::findOrFail($id);
+        $this->authorize('destroy', $answer);
         try {
-            $answer = Answer::where('id', $id)->firstOrFail();
             $answer->delete();
             return redirect('questions/' . $answer->id_question);
         }
