+++ app/Http/Controllers/QuestionController.php
@@ -19,16 +19,28 @@ class QuestionController extends Controller
 
     public function showMostLikedQuestions(Request $request)
     {
-        // exact search
-        if ($request->has('text')) {
+        if ($request->has('text') && $request->get('text') != '') {
             $searchTerm = $request->get('text');
 
-            $questions = Question::where('title', 'ILIKE', '%' . $searchTerm . '%')
-                ->orWhere('content', 'ILIKE', '%' . $searchTerm . '%')
-                ->with(['user', 'community', 'likes', 'dislikes', 'answers'])
+            // check for exact match (enclosed in quotes)
+            if (preg_match('/^".+"$/', $searchTerm)) {
+                $searchTerm = trim($searchTerm, '"');
+
+                $questions = Question::where('title', 'ILIKE', '%' . $searchTerm . '%')
+                    ->orWhere('content', 'ILIKE', '%' . $searchTerm . '%');
+
+            } else {
+                // perform full text search
+                $formattedTerm = str_replace(' ', ' | ', $searchTerm);
+
+                $questions = Question::whereRaw("tsvectors @@ to_tsquery('english', ?)", [$formattedTerm]);
+            }
+
+            $questions = $questions->with(['user', 'community', 'likes', 'dislikes', 'answers'])
                 ->withCount(['likes', 'dislikes', 'answers'])
                 ->orderBy('likes_count', 'desc')
                 ->get();
+
         } else {
 
             $questions = Question::with(['user', 'community', 'likes', 'dislikes', 'answers'])
@@ -62,13 +74,14 @@ class QuestionController extends Controller
      */
     public function show(int $id)
     {
+        $this->authorize('show', Question::class);
         try {
 
             $answers = Answer::with(['user', 'likes', 'dislikes'])->where('id_question', $id)
                 ->get();
             return view('questions.show', [
-                'question' => Question::where('id', $id)->firstOrFail(),
-                'answers' => $answers            
+                'question' => Question::findOrFail($id),
+                'answers' => $answers
             ]);
         } catch (ModelNotFoundException $e) {
             return "Question not found.";
@@ -80,10 +93,10 @@ class QuestionController extends Controller
      */
     public function edit(int $id)
     {
+        $question = Question::findOrFail($id);
+        $this->authorize('edit', $question);
         try {
-            return view('questions.edit', [
-                'question' => Question::where('id', $id)->firstOrFail()
-            ]);
+            return view('questions.edit', ['question' => $question]);
         } catch (ModelNotFoundException $e) {
             return "Question not found.";
         }
@@ -94,8 +107,9 @@ class QuestionController extends Controller
      */
     public function update(Request $request, int $id)
     {
+        $question = Question::findOrFail($id);
+        $this->authorize('update', $question);
         try {
-            $question = Question::where('id', $id)->firstOrFail();
             $question->title = $request->input('title');
             $question->content = $request->input('content');
             $question->save();
@@ -110,12 +124,33 @@ class QuestionController extends Controller
      */
     public function destroy(int $id)
     {
+        $question = Question::findOrFail($id);
+        $this->authorize('destroy', $question);
         try {
-            $question = Question::where('id', $id)->firstOrFail();
             $question->delete();
             return redirect('questions/');
         } catch (ModelNotFoundException $e) {
             return "Question not found.";
         }
+    }
+
+    public function postQuestion(Request $request){
+        $validatedData = $request->validate([
+            'title' => 'required|string|max:255',
+            'content' => 'required|string',
+            'id_user' => 'required|integer',
+            'id_community' => 'required|integer',
+        ]);
+
+        $question = new Question;
+        $question->title = $validatedData['title'];
+        $question->content = $validatedData['content'];
+        $question->id_user = $validatedData['id_user'];
+        $question->id_community = $validatedData['id_community'];
+
+        $question->save();
+
+        return redirect()->route('questions' )
+        ->withSuccess('Question posted successfully!');
     }
 }
